+++ src/lib/utility/scheduling/TaskGroupSequence.cpp
@@ -17,7 +17,7 @@ Task::TaskState TaskGroupSequence::doUpdate(std::shared_ptr<Blackboard> blackboa
     return STATE_FAILURE;
   }
 
-  TaskState state = m_taskRunners[m_taskIndex]->update(blackboard);
+  TaskState state = m_taskRunners[static_cast<size_t>(m_taskIndex)]->update(blackboard);
 
   if(state == STATE_SUCCESS) {
     m_taskIndex++;
